removes duplicate keypairsvar, adds test for native kepairs
