+++ src/circuit/vanchor/constraints.rs
@@ -1,20 +1,46 @@
 use ark_crypto_primitives::{crh::CRHGadget, CRH};
 use ark_ff::fields::PrimeField;
-use ark_r1cs_std::prelude::*;
+use ark_r1cs_std::{fields::fp::FpVar, prelude::*};
 
 use ark_std::marker::PhantomData;
 
+use crate::leaf::{NewLeafCreation, NewLeafCreationGadget};
+
 #[derive(Clone)]
-pub struct KeypairsVar<F: PrimeField, H: CRH, HG: CRHGadget<H, F>> {
+pub struct KeypairsVar<
+	H: CRH,
+	HG: CRHGadget<H, F>,
+	L: NewLeafCreation<H>,
+	LG: NewLeafCreationGadget<F, H, HG, L>,
+	F: PrimeField,
+> {
 	pubkey_var: <HG as CRHGadget<H, F>>::OutputVar,
-	data: PhantomData<F>,
+	privkey_var: FpVar<F>,
+	_leaf_creation: PhantomData<L>,
+	_leaf_creation_gadget: PhantomData<LG>,
 }
-impl<F: PrimeField, H: CRH, HG: CRHGadget<H, F>> KeypairsVar<F, H, HG> {
-	pub fn compute_public_key(h: &HG::ParametersVar, privkey: &Vec<UInt8<F>>) -> Self {
-		let pubkey_var = HG::evaluate(&h, &privkey).unwrap();
+
+impl<
+		H: CRH,
+		HG: CRHGadget<H, F>,
+		L: NewLeafCreation<H>,
+		LG: NewLeafCreationGadget<F, H, HG, L>,
+		F: PrimeField,
+	> KeypairsVar<H, HG, L, LG, F>
+{
+	pub fn public_key_var(
+		h: HG::ParametersVar,
+		secrets: <LG as NewLeafCreationGadget<F, H, HG, L>>::PrivateVar,
+	) -> Self {
+		let privkey_var = LG::get_privat_key(&secrets).unwrap();
+		let mut bytes = Vec::<UInt8<F>>::new();
+		bytes.extend(privkey_var.to_bytes().unwrap());
+		let pubkey_var = HG::evaluate(&h, &bytes).unwrap();
 		KeypairsVar {
 			pubkey_var,
-			data: PhantomData,
+			privkey_var,
+			_leaf_creation: PhantomData,
+			_leaf_creation_gadget: PhantomData,
 		}
 	}
 }
@@ -24,7 +50,10 @@ impl<F: PrimeField, H: CRH, HG: CRHGadget<H, F>> KeypairsVar<F, H, HG> {
 mod test {
 	use super::*;
 	use crate::{
-		leaf::{newleaf::NewLeaf, NewLeafCreation},
+		leaf::newleaf::{
+			constraints::{NewLeafGadget, PrivateVar},
+			NewLeaf,
+		},
 		poseidon::{
 			constraints::{CRHGadget, PoseidonParametersVar},
 			sbox::PoseidonSbox,
@@ -35,9 +64,13 @@ mod test {
 	use ark_bls12_381::Fq;
 	use ark_crypto_primitives::crh::{constraints::CRHGadget as CRHGadgetTrait, CRH as CRHTrait};
 	use ark_ff::to_bytes;
+	use ark_r1cs_std::{
+		alloc::{AllocVar, AllocationMode},
+		prelude::EqGadget,
+		R1CSVar,
+	};
 	use ark_relations::r1cs::ConstraintSystem;
 	use ark_std::test_rng;
-	use KeypairsVar;
 	#[derive(Default, Clone)]
 	struct PoseidonRounds3;
 
@@ -52,11 +85,9 @@ mod test {
 	type PoseidonCRH3Gadget = CRHGadget<Fq, PoseidonRounds3>;
 
 	type Leaf = NewLeaf<Fq, PoseidonCRH3>;
-
-	type Keypairs = KeypairsVar<Fq, PoseidonCRH3, PoseidonCRH3Gadget>;
-
+	type LeafGadget = NewLeafGadget<Fq, PoseidonCRH3, PoseidonCRH3Gadget, Leaf>;
 	#[test]
-	fn should_crate_publik_key_constraints() {
+	fn should_crate_new_public_key() {
 		let rng = &mut test_rng();
 		let cs = ConstraintSystem::<Fq>::new_ref();
 
@@ -69,17 +100,24 @@ mod test {
 		let pubkey = PoseidonCRH3::evaluate(&params, &privkey).unwrap();
 
 		// Constraints version
+		let secrets_var = PrivateVar::new_witness(cs.clone(), || Ok(&secrets)).unwrap();
 		let bytes = to_bytes![secrets.priv_key].unwrap();
 		let privkey_var = Vec::<UInt8<Fq>>::new_witness(cs.clone(), || Ok(bytes)).unwrap();
 		let params_var =
 			PoseidonParametersVar::new_variable(cs, || Ok(&params), AllocationMode::Constant)
 				.unwrap();
 		let pubkey_var = PoseidonCRH3Gadget::evaluate(&params_var, &privkey_var).unwrap();
-		let keypair = Keypairs::compute_public_key(&params_var, &privkey_var);
-		let new_pubkey_var = keypair.pubkey_var.value().unwrap();
+		let keypairs =
+			KeypairsVar::<PoseidonCRH3, PoseidonCRH3Gadget, Leaf, LeafGadget, Fq>::public_key_var(
+				params_var,
+				secrets_var,
+			);
+		let new_pubkey_var = keypairs.pubkey_var;
+		let res = pubkey_var.is_eq(&new_pubkey_var).unwrap();
+		
 		// Check equality
-		let res = pubkey_var.value().unwrap();
-		assert_eq!(res, pubkey);
-		assert_eq!(res, new_pubkey_var);
+		assert!(res.value().unwrap());
+		assert_eq!(pubkey, new_pubkey_var.value().unwrap());
+		assert!(res.cs().is_satisfied().unwrap());
 	}
 }
